Frontend data is a mess

+++ api/src/main/java/io/trenchcrusade/api/rule/equipment/Equipment.java
@@ -11,8 +11,8 @@ import java.util.Set;
 public class Equipment {
     @Id // KEY
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
-    public Integer getId() {
+    private Long id;
+    public Long getId() {
         return id;
     }
