+++ api/src/main/java/io/trenchcrusade/api/rule/equipment/EquipmentRepository.java
@@ -2,4 +2,4 @@ package io.trenchcrusade.api.rule.equipment;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface EquipmentRepository extends CrudRepository<Equipment, Integer> {}
+public interface EquipmentRepository extends CrudRepository<Equipment, Long> {}
