+++ api/src/main/java/io/trenchcrusade/api/rule/faction/FactionRepository.java
@@ -2,4 +2,4 @@ package io.trenchcrusade.api.rule.faction;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface FactionRepository extends CrudRepository<Faction, Integer> {}
+public interface FactionRepository extends CrudRepository<Faction, Long> {}
