+++ api/src/main/java/io/trenchcrusade/api/rule/faction/FactionTroopType.java
@@ -7,8 +7,8 @@ import jakarta.persistence.*;
 public class FactionTroopType {
     @Id // KEY
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-    public Integer getId() {
+    private Long id;
+    public Long getId() {
         return id;
     }
