+++ api/src/main/java/io/trenchcrusade/api/rule/keyword/KeywordRepository.java
@@ -2,4 +2,4 @@ package io.trenchcrusade.api.rule.keyword;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface KeywordRepository extends CrudRepository<Keyword, Integer> {}
+public interface KeywordRepository extends CrudRepository<Keyword, Long> {}
