+++ api/src/main/java/io/trenchcrusade/api/rule/troop_type/TroopType.java
@@ -12,8 +12,8 @@ import java.util.Set;
 public class TroopType {
     @Id // KEY
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
-    public Integer getId() {
+    private Long id;
+    public Long getId() {
         return id;
     }
