+++ api/src/main/java/io/trenchcrusade/api/rule/troop_type/TroopTypeRepository.java
@@ -2,4 +2,4 @@ package io.trenchcrusade.api.rule.troop_type;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface TroopTypeRepository extends CrudRepository<TroopType, Integer> {}
+public interface TroopTypeRepository extends CrudRepository<TroopType, Long> {}
