+++ api/src/main/java/io/trenchcrusade/api/rule/troop_type/TroopTypeRule.java
@@ -6,8 +6,8 @@ import jakarta.persistence.*;
 public class TroopTypeRule {
     @Id // KEY
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-    public Integer getId() {
+    private Long id;
+    public Long getId() {
         return id;
     }
